frontend setup for reputation

+++ db/migrate/20210520104141_create_ratings.rb
@@ -3,7 +3,7 @@ class CreateRatings < ActiveRecord::Migration[5.2]
     create_table :ratings do |t|
       t.integer :challenge_leaderboard_extra_id, null: false
       t.integer :participant_id, null: false
-      t.integer :rating, null: false
+      t.float   :rating, null: false
       t.float   :mu
       t.float   :sigma
       t.integer :rank
